fix: fix the problem of button with href prop
There would be problems when href prop of button is been set.
Problem1: the button would become a link even the href's value is blank white space
Problem2: if the user click on the button, there would be a nested low code editor shown in the current low code editor.